Refactor package Homura to homura and update build.gradle
