+++ src/main/java/homura/EmptyInputHomuraException.java
@@ -6,7 +6,7 @@
 // https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
 // and
 // https://www.w3schools.com/java/java_packages.asp
-package Homura;
+package homura;
 
 public class EmptyInputHomuraException extends HomuraRuntimeException {
     public static final String indent = "    ";
