+++ src/main/java/homura/HomuraRuntimeException.java
@@ -6,7 +6,7 @@
 // https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
 // and
 // https://www.w3schools.com/java/java_packages.asp
-package Homura;
+package homura;
 
 public class HomuraRuntimeException extends RuntimeException {
     /**
