+++ src/main/java/homura/HomuraUtils.java
@@ -8,7 +8,7 @@
 // https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
 // and
 // https://www.w3schools.com/java/java_packages.asp
-package Homura;
+package homura;
 
 import java.io.File;
 import java.io.FileWriter;
