+++ src/main/java/homura/InvalidCmdHomuraException.java
@@ -1,29 +1,26 @@
 /**
- * An exception for when a command receives an invalid input.
+ * An exception for when an invalid command is passed to Homura.
  */
 
 // How to use and create packages inspired by
 // https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
 // and
 // https://www.w3schools.com/java/java_packages.asp
-package Homura;
+package homura;
 
-public class InvalidInputHomuraException extends HomuraRuntimeException {
+public class InvalidCmdHomuraException extends HomuraRuntimeException {
     public static final String indent = "    ";
 
     private String cmd;
-    private String inp;
 
     /**
-     * Creates an InvalidInputHomuraException.
-     *
-     * @param cmd The passed command.
-     * @param inp The full passed input.
+     * Creates an InvalidCmdHomuraException.
+     * 
+     * @param cmd The invalid passed command.
      */
-    public InvalidInputHomuraException(String cmd, String inp) {
+    public InvalidCmdHomuraException(String cmd) {
         super();
         this.cmd = cmd;
-        this.inp = inp;
     }
 
     /**
@@ -34,7 +31,6 @@ public class InvalidInputHomuraException extends HomuraRuntimeException {
     @Override
     public String toString() {
         return getClass().getSimpleName()
-            + " - Invalid input passed to cmd " + cmd + ":" + '\n'
-            + indent + indent + inp;
+            + " - Invalid cmd passed: " + cmd;
     }
 }
