+++ src/main/java/homura/Playground.java
@@ -5,7 +5,7 @@
 // https://github.com/weiseng18/ip/tree/master/src/main/java/panorama
 // and
 // https://www.w3schools.com/java/java_packages.asp
-package Homura;
+package homura;
 
 
 import java.util.*;
