+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-     Hi, I'm Akemi Homura.
+     Hi, I'm Akemi homura.
      Have you seem Madoka anywhere?
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
